tests: major refactor
