--- packages/integration-tests/test/setup.ts
@@ -1,6 +1,0 @@
-/* External Imports */
-import chai = require('chai')
-import chaiAsPromised = require('chai-as-promised')
-
-chai.use(chaiAsPromised)
-const should = chai.should()
